Moved core js instrumentation methods to a separate file
